+++ cmd/mdmctl/get_users.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/micromdm/micromdm/platform/apiserver/list"
+	"github.com/micromdm/micromdm/platform/api/server/list"
 )
 
 type usersTableOutput struct{ w *tabwriter.Writer }
