+++ cmd/micromdm/serve.go
@@ -43,9 +43,9 @@ import (
 	"github.com/micromdm/micromdm/mdm/connect"
 	"github.com/micromdm/micromdm/mdm/enroll"
 	"github.com/micromdm/micromdm/pkg/crypto"
-	"github.com/micromdm/micromdm/platform/apiserver/apply"
-	"github.com/micromdm/micromdm/platform/apiserver/list"
-	"github.com/micromdm/micromdm/platform/apiserver/remove"
+	"github.com/micromdm/micromdm/platform/api/server/apply"
+	"github.com/micromdm/micromdm/platform/api/server/list"
+	"github.com/micromdm/micromdm/platform/api/server/remove"
 	"github.com/micromdm/micromdm/platform/apns"
 	"github.com/micromdm/micromdm/platform/appstore"
 	"github.com/micromdm/micromdm/platform/blueprint"
