+++ platform/api/server/list/endpoint.go
