+++ platform/api/server/list/service.go
