+++ platform/api/server/remove/client.go
