+++ platform/api/server/remove/endpoint.go
