+++ platform/api/server/remove/service.go
