+++ platform/api/server/remove/transport_http.go
